requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
gTTS>=2.3.2
Flask>=3.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.0
reportlab>=4.0.0
deep-translator>=1.11.4
opencv-python-headless>=4.8.0
numpy>=1.26.0
openai>=1.40.0
python-dotenv>=1.0.1
gunicorn>=22.0.0
